moved types and enums to type file rather than lib file
