+++ app/src/components/proposals/proposal.types.ts
@@ -3,20 +3,101 @@ import React from "react"
 
 import { CardActionsProps } from "@/components/proposals/proposalPage/VotingSidebar/votingCard/VoterActions"
 import { CitizenshipQualification } from "@/lib/types"
+import { ProposalBadgeType } from "@/components/proposals/proposalsPage/components/ProposalCard"
 
-export interface ProposalPageDataInterface {
-  user: User | null
-  citizen?: Citizen
-  votingOpen: boolean
-  votingComplete: boolean
-  voted: boolean
-  votingRecord?: VotingRecordInterface
-  startDate: Date
-  endDate: Date
+export type OffChainProposalResponse = {
+  meta: {
+    has_next: boolean
+    total_returned: number
+    next_offset: number
+  }
+  data: OffChainProposal[]
+}
+
+export enum ProposalStatus {
+  PENDING = "PENDING",
+  ACTIVE = "ACTIVE",
+  CANCELLED = "CANCELLED",
+  EXECUTED = "EXECUTED",
+  QUEUED = "QUEUED",
+  FAILED = "FAILED",
+  DEFEATED = "DEFEATED",
+}
+
+export type ProposalData = {
+  id: string
+  proposer: string
+  snapshotBlockNumber: number
+  createdTime: string
+  startTime: string
+  startBlock?: string
+  endTime: string
+  endBlock?: string
+  cancelledTime?: string | null
+  executedTime?: string | null
+  executedBlock?: string | null
+  queuedTime?: string | null
+  markdowntitle: string
+  description: string
+  quorum: string
+  approvalThreshold?: string
+  proposalData: object
+  unformattedProposalData?: string | null
+  proposalResults: object
   proposalType: ProposalType
-  proposalId: string
-  proposalStatus: string
-  citizenEligibility: CitizenshipQualification | null
+  status: ProposalStatus
+  createdTransactionHash?: string | null
+  cancelledTransactionHash?: string | null
+  executedTransactionHash?: string | null
+  proposalTemplate?: object
+  // This value should always be included in offchain and hybrid proposals
+  offchainProposalId: string
+}
+
+export type OffChainProposal = {
+  id: string
+  proposer: string
+  snapshotBlockNumber: number
+  createdTime: string
+  startTime: string
+  startBlock: string
+  endTime: string
+  endBlock: string
+  cancelledTime: string | null
+  executedTime: string | null
+  executedBlock: string | null
+  queuedTime: string | null
+  markdowntitle: string
+  description: string
+  quorum: string
+  proposalData: object // We can define this more specifically if needed
+  proposalResults: object // We can define this more specifically if needed
+  proposalType: ProposalType
+  status: ProposalStatus
+  offchainProposalId: string
+}
+
+// For UI
+export type UIProposal = {
+  id: string
+  badge: {
+    badgeType: ProposalBadgeType
+  }
+  voted?: boolean
+  passed?: boolean
+  textContent: {
+    title: string
+    subtitle?: string
+  }
+  dates: {
+    startDate: string
+    endDate: string
+  }
+  arrow: {
+    href: string
+  }
+  proposalResults?: object
+  proposalType: ProposalType
 }
 
 export interface VotingRecordInterface {
