+++ app/src/lib/utils/voting.ts
@@ -1,11 +1,11 @@
 import { JsonValue } from "@prisma/client/runtime/library"
 
 import {
+  ProposalData,
   ProposalType,
   VoteType,
   VotingCardProps,
 } from "@/components/proposals/proposal.types"
-import { ProposalData } from "@/lib/proposals"
 import { CitizenshipQualification } from "@/lib/types"
 
 const API_URL = process.env.NEXT_PUBLIC_AGORA_API_URL
